Wow! All that work in one commit?! 😎
Just kidding. I did all the work in the past. Just moving it to GitHub. 

Note: The reason why there are copies of number 49 is because I just change the data. Makes it slightly easier than copy+pasting a bunch.... just ignore them and all the errors they make. 👍

Also: NO AI WROTE ANY OF THIS. Actually, there is like one sentence that when I had the AI check for spelling errors, it decided to rewrite like one or two sentences. I was tired and didn't change it at the time. 🤷‍♂️